% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{functionTimer}
\alias{functionTimer}
\title{functionTimer}
\usage{
functionTimer(reps = 1)
}
\arguments{
\item{reps}{number of times to repeat block}
}
\description{
Calls an Rcpp function that includes a timer around an editable block. Useful for comparing speed of different function implementations.
}
