% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{loadShapeFile}
\alias{loadShapeFile}
\title{loadShapeFile}
\usage{
loadShapeFile(name = "DRC_admin1")
}
\description{
Loads a shape file (i.e. a SpatialPolygonsDataFrame) with a given name from the inst/extdata folder.
}
