% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{merge.SpatialPolygonsDataFrame}
\alias{merge.SpatialPolygonsDataFrame}
\title{merge.SpatialPolygonsDataFrame}
\usage{
\method{merge}{SpatialPolygonsDataFrame}(shp, df)
}
\description{
An ordinary merge operation sometimes changes the order of rows, which when applied to SpatialPolygonsDataFrame data can mean data rows no longer correspond to the correct polygon. This function merges data while avoiding this issue.
}
