% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{rateRatio}
\alias{rateRatio}
\title{rateRatio}
\usage{
rateRatio(count1, time1, count2, time2, alpha = 0.05)
}
\description{
Computes point estimate and upper and lower confidence intervals on a ratio of rates. Default method is to enter raw counts and time periods, but if entering rates simply set time1=1 and time2=1.
}
