% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmental.R
\name{simQuantiles}
\alias{simQuantiles}
\title{simQuantiles}
\usage{
simQuantiles(FUN = "SIS_stochastic_hybrid", args = list(), reps = 100,
  quantiles = c(0.05, 0.5, 0.95))
}
\arguments{
\item{FUN}{the stochastic simulation function to use.}

\item{args}{a list of arguments to the function.}

\item{reps}{number of times to repeat the stochastic simulation.}

\item{quantiles}{which quantiles to compute over replicates.}
}
\description{
Runs a given stochastic simulation function many times, computing the mean and quantiles over replicates. Note that this method will only work with simulations that have a fixed time step, i.e. synchronous or hybrid simulations, and not with asynchronous simulations. In the hybrid case the maxIterations limit cannot be reached in any simulation.
}
