% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmental.R
\name{SIS_analytical}
\alias{SIS_analytical}
\title{SIS_analytical}
\usage{
SIS_analytical(beta = 1, r = 0.25, I_init = 10, N = 1000,
  times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Returns analytical solution to deterministic SIS model. At equilibrium the number of infectives is given by \eqn{I* = N(1 - r/beta)}.
}
