% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmental.R
\name{SIS_stochastic_async}
\alias{SIS_stochastic_async}
\title{SIS_stochastic_async}
\usage{
SIS_stochastic_async(beta = 1, r = 0.25, I_init = 100, N = 1000,
  maxIterations = 10000)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{N}{total number of individuals in population.}

\item{maxIterations}{exit if this number of iterations is reached.}
}
\description{
Draw from asynchronous stochastic SIS model. Return state of the system at all time points at which any event occurs. Stop when maxIterations is reached.
}
